### 2026-04-01 b67ac49
Add MIT License to the project
