### 2026-04-01 d13c46b
Add required JCON paper sections (Statement of need, State of the field, Software design, Research impact statement, AI usage disclosure)
Added sections on the statement of need, state of the field, software design, research impact statement, and AI usage disclosure to enhance the manuscript's comprehensiveness and clarity.
