+++ paper/paper.tex
@@ -45,6 +45,17 @@ Despite recent advances in Scientific Machine Learning applications to biologica
 
 Our work addresses these limitations through a progressive research workflow with several key contributions: (i) the development of a dual-network hybrid UDE as an initial exploration that augments mechanistic growth with learned immune dynamics, (ii) the evolution to a pure mechanistic UDE that builds upon hybrid insights while maintaining full parameter interpretability, (iii) implementation of rigorous cross-validation with validation-based early stopping to ensure robust model evaluation, (iv) comprehensive forecasting analysis demonstrating superior predictive accuracy for both short-term and long-term horizons, (v) implementation of biologically consistent constraints and physics-informed priors, and (vi) provision of a fully reproducible Julia implementation with extensive performance metrics.
 
+
+% ===========================
+\section{Statement of need}
+
+Computational oncology lacks accessible, interpretable tools for modeling tumor--immune interactions that balance mechanistic understanding with predictive accuracy. Existing approaches are either purely mechanistic models that underfit complex biological dynamics, or black-box machine learning methods that lack biological interpretability. Clinicians and researchers need models that can provide both accurate forecasts and mechanistic insight into tumor--immune dynamics. Our Julia package addresses this gap by providing a UDE-based framework for tumor--immune modeling that: (i) integrates domain knowledge through physics-informed constraints, (ii) learns interpretable biological parameters within classical mathematical frameworks, (iii) provides robust forecasting capabilities validated through cross-validation, and (iv) leverages Julia's SciML ecosystem for computational efficiency, making these methods practical for research and potential clinical translation.
+
+% ===========================
+\section{State of the field}
+
+Several approaches exist for modeling tumor--immune dynamics. Classical mechanistic models such as the Kuznetsov--Taylor model \cite{kuznetsov1994} and the de Pillis--Radunskaya system \cite{depillis2005validated} provide interpretable frameworks but often become parameter-heavy and difficult to calibrate with limited clinical data. The Gompertz growth model \cite{vaghi2020gompertz} offers a reliable description of tumor growth kinetics but lacks immune interaction terms. On the machine learning side, Neural ODEs \cite{chen2018node} and Physics-Informed Neural Networks \cite{raissi2019pinn} have shown promise in scientific applications, but their direct application to tumor--immune modeling remains limited. Universal Differential Equations \cite{rackauckas2020ude} represent the most relevant advancement, enabling hybrid mechanistic-neural models. However, existing UDE applications in biology \cite{shen2023differentiable,aboelyazeed2023differentiable,dandekar2020machine} have not specifically addressed tumor--immune dynamics with rigorous cross-validation and forecasting frameworks. Our work distinguishes itself by: (i) applying UDEs specifically to tumor--immune interactions with biologically-constrained Gompertz growth and Michaelis--Menten immune killing, (ii) implementing a progressive workflow from hybrid to pure mechanistic UDEs that achieves both high accuracy and full interpretability, and (iii) providing comprehensive cross-validation and forecasting evaluation suitable for clinical translation assessment.
+
 % ===========================
 \section{Methodology}
 
@@ -311,6 +322,25 @@ Given sparse immune measurements, we construct continuous immune functions $I(t)
 
 \subsubsection{Group Formation}
 Data is grouped by \{KineticID, TumorID\} combinations, yielding independent trajectories for model training and evaluation. Groups with fewer than 2 time points are excluded to ensure adequate temporal resolution for forecasting evaluation.
+
+% ===========================
+\section{Software design}
+
+Our implementation is organized as a modular Julia package leveraging the SciML ecosystem. The codebase follows a clear separation of concerns with the following structure:
+
[…]
