### 2026-04-01 0634751
Fix invalid DOI and add missing DOIs in ref.bib
Added DOI fields to several references and corrected some entries.
