+++ paper/ref.bib
@@ -138,7 +138,8 @@
   author    = {Dupont, Emilien and Doucet, Arnaud and Teh, Yee Whye},
   booktitle = {Advances in Neural Information Processing Systems},
   volume    = {32},
-  year      = {2019}
+  year      = {2019},
+  doi          = {10.48550/arXiv.1904.01681}
 }
 
 @article{massaroli2020dissecting,
@@ -147,7 +148,8 @@
   journal = {Advances in Neural Information Processing Systems},
   volume  = {33},
   pages   = {3952--3963},
-  year    = {2020}
+  year    = {2020},
+  doi          = {10.48550/arXiv.2002.08071}
 }
 
 @article{finlay2020train,
@@ -156,7 +158,8 @@
   year          = {2020},
   eprint        = {2002.02798},
   archivePrefix = {arXiv},
-  primaryClass  = {cs.LG}
+  primaryClass  = {cs.LG},
+  doi          = {10.48550/arXiv.2002.02798}
 }
 
 @article{kim2021stiff,
@@ -188,7 +191,7 @@
   volume  = {240},
   pages   = {111992},
   year    = {2022},
-  doi     = {10.1016/j.combustflame.2021.111992}
+  doi     = {10.1016/j.combustflame.2022.111992}
 }
 
 @article{lai2021structural,
@@ -232,6 +235,243 @@
   title   = {Differentiable modelling to unify machine learning and physical models for geosciences},
   author  = {Shen, Chaopeng and Appling, Alison P. and Gentine, Pierre and Bandai, Toshiyuki and Gupta, Hoshin and Tartakovsky, Alexandre and Baity-Jesi, Marco and Fenicia, Fabrizio and Kifer, Daniel and Li, Li and others},
   journal = {Nature Reviews Earth \& Environment},
+  title        = {Neural Ordinary Differential Equations},
+  author       = {Chen, Ricky T.~Q. and Rubanova, Yulia and Bettencourt, Jesse and Duvenaud, David},
+  booktitle    = {Advances in Neural Information Processing Systems (NeurIPS)},
+  year         = {2018},
+  doi          = {10.48550/arXiv.1806.07366}
+}
+
+@article{innes2018flux,
+  title   = {Flux: Elegant machine learning with {J}ulia},
+  author  = {Innes, Mike},
+  journal = {Journal of Open Source Software},
+  volume  = {3},
+  number  = {25},
+  pages   = {602},
+  year    = {2018},
+  doi     = {10.21105/joss.00602}
+}
+
+@article{rackauckas2017de,
+  title   = {{D}ifferential{E}quations.jl -- a performant and feature-rich ecosystem for solving differential equations in {J}ulia},
+  author  = {Rackauckas, Chris and Nie, Qing},
+  journal = {Journal of Open Research Software},
+  volume  = {5},
+  number  = {1},
+  pages   = {15},
+  year    = {2017},
+  doi     = {10.5334/jors.151}
+}
+
+@misc{rackauckas2019diffeqflux,
+  title         = {{D}iff{E}q{F}lux.jl -- A Julia library for neural differential equations},
+  author        = {Rackauckas, Chris and Innes, Mike and Ma, Yingbo and others},
+  year          = {2019},
+  eprint        = {1902.02376},
+  archivePrefix = {arXiv},
+  primaryClass  = {cs.LG},
+  doi           = {10.48550/arXiv.1902.02376}
+}
+
+@article{rackauckas2020ude,
+  title         = {Universal Differential Equations for Scientific Machine Learning},
[…]
