### 2026-04-01 22a2bc3
Fix ref.bib: remove duplicate entries, add missing DOIs, fix invalid DOI
