+++ paper/ref.bib
@@ -4,7 +4,7 @@
   booktitle    = {Advances in Neural Information Processing Systems (NeurIPS)},
   year         = {2018},
   doi          = {10.48550/arXiv.1806.07366}
-}
+} 
 
 @article{innes2018flux,
   title   = {Flux: Elegant machine learning with {J}ulia},
