### 2026-04-01 ee73114
Fix ref.bib: remove duplicates, fix invalid DOI, add 6 missing DOIs
Removed several outdated references from the bibliography file.
