+++ paper/ref.bib
@@ -4,242 +4,6 @@
   booktitle    = {Advances in Neural Information Processing Systems (NeurIPS)},
   year         = {2018},
   doi          = {10.48550/arXiv.1806.07366}
-} 
-
-@article{innes2018flux,
-  title   = {Flux: Elegant machine learning with {J}ulia},
-  author  = {Innes, Mike},
-  journal = {Journal of Open Source Software},
-  volume  = {3},
-  number  = {25},
-  pages   = {602},
-  year    = {2018},
-  doi     = {10.21105/joss.00602}
-}
-
-@article{rackauckas2017de,
-  title   = {{D}ifferential{E}quations.jl -- a performant and feature-rich ecosystem for solving differential equations in {J}ulia},
-  author  = {Rackauckas, Chris and Nie, Qing},
-  journal = {Journal of Open Research Software},
-  volume  = {5},
-  number  = {1},
-  pages   = {15},
-  year    = {2017},
-  doi     = {10.5334/jors.151}
-}
-
-@misc{rackauckas2019diffeqflux,
-  title         = {{D}iff{E}q{F}lux.jl -- A Julia library for neural differential equations},
-  author        = {Rackauckas, Chris and Innes, Mike and Ma, Yingbo and others},
-  year          = {2019},
-  eprint        = {1902.02376},
-  archivePrefix = {arXiv},
-  primaryClass  = {cs.LG},
-  doi           = {10.48550/arXiv.1902.02376}
-}
-
-@article{rackauckas2020ude,
-  title         = {Universal Differential Equations for Scientific Machine Learning},
-  author        = {Rackauckas, Christopher and Ma, Yingbo and Martensen, Julius and Warner, Collin and Zubov, Kirill and Supekar, Rohit and Skinner, Dominic and Ramadhan, Ali and Edelman, Alan},
-  year          = {2020},
-  eprint        = {2001.04385},
-  archivePrefix = {arXiv},
-  primaryClass  = {cs.LG},
-  doi           = {10.48550/arXiv.2001.04385}
-}
-
-@article{kuznetsov1994,
-  title   = {Nonlinear dynamics of immunogenic tumors: parameter estimation and global bifurcation analysis},
-  author  = {Kuznetsov, Vladimir A. and Makalkin, I. A. and Taylor, M. A. and Perelson, Alan S.},
-  journal = {Bulletin of Mathematical Biology},
-  volume  = {56},
-  number  = {2},
-  pages   = {295--321},
-  year    = {1994},
-  doi     = {10.1007/BF02460644}
-}
-
-@article{depillis2005validated,
-  title   = {A validated mathematical model of cell-mediated immune response to tumor growth},
-  author  = {de Pillis, Lisette G. and Radunskaya, Ami E. and Wiseman, Charles L.},
-  journal = {Cancer Research},
-  volume  = {65},
-  number  = {17},
-  pages   = {7950--7958},
-  year    = {2005},
-  doi     = {10.1158/0008-5472.CAN-05-0564}
-}
-
-@article{vaghi2020gompertz,
-  title   = {Population modeling of tumor growth curves and the reduced Gompertz model improve prediction of the age of experimental tumors},
-  author  = {Vaghi, Cristina and Rodallec, Alix and Fanciullino, Raphaelle and Ciccolini, Joseph and Mochel, Jonathan P. and Mastri, Michalis and Poignard, Clair and Ebos, John M. L. and Benzekry, Sebastien},
-  journal = {PLOS Computational Biology},
-  volume  = {16},
-  number  = {2},
-  pages   = {e1007178},
-  year    = {2020},
-  doi     = {10.1371/journal.pcbi.1007178}
-}
-
-@article{raissi2019pinn,
[…]
