### 2026-04-07 03a12f2
Rewrite paper.tex: address reviewer critique, reduce verbosity, add limitations
- Rename 'Pure Mechanistic UDE' to 'Structured-Parameter UDE' with terminology note
- Remove overclaiming/clinical framing language
- Add comprehensive Limitations section (baselines, ablation, UQ, naming, dataset scope, forecasting)
- Acknowledge weak baselines in Results and table caption
- Cut from ~12 pages to ~8-10 pages by removing redundant sections
- Note overfitting gap in Discussion
