+++ paper/paper.tex
@@ -14,8 +14,8 @@
 
 \graphicspath{{figures/}}
 
-\title{Biologically Consistent Universal Differential Equations for Tumor--Immune Dynamics:\\
-Towards a more interpretable approach}
+\title{Universal Differential Equations for Tumor--Immune Dynamics:\\
+A Progressive Modeling Workflow in Julia's SciML Ecosystem}
 
 \author[1]{Anish Sarkar\thanks{ORCID: \href{https://orcid.org/0009-0009-7965-7788}{0009-0009-7965-7788}}}
 \affil[1]{Netaji Subhash Engineering College (MAKAUT)}
@@ -25,329 +25,137 @@ Towards a more interpretable approach}
 \maketitle
 
 \begin{abstract}
-We present a comprehensive study of Universal Differential Equations (UDEs) for modeling tumor--immune dynamics, implemented using Julia's SciML ecosystem. We developed and compared two UDE architectures in a progressive workflow: \textbf{(A) Dual-network hybrid approach} that initially augments mechanistic Gompertz growth with neural networks to capture unmodeled immune dynamics and temporal corrections, achieving $R^2 = 0.897$, and \textbf{(B) Pure mechanistic UDE approach} that builds upon the hybrid insights to learn interpretable scientific parameters within classical Gompertz and Michaelis--Menten frameworks, achieving superior accuracy ($R^2 = 0.991$) with enhanced parameter interpretability. Both variants enforce biological constraints through physics-informed loss functions and consistency priors. Using cross-validation with validation-based early stopping, we demonstrate robust forecasting capabilities with Mean Absolute Percentage Error (MAPE) of 1.8\% for short-term and 4.3\% for medium-term predictions. Our implementation leverages Julia's differentiable programming capabilities through DifferentialEquations.jl, Flux.jl, and SciMLSensitivity.jl, achieving significant computational efficiency gains. This progressive approach advances the application of Scientific Machine Learning in oncology by providing interpretable, physics-constrained models for tumor--immune interactions with demonstrated forecasting accuracy and clinical translation potential.
+We present a study of Universal Differential Equations (UDEs) for modeling tumor--immune dynamics using Julia's SciML ecosystem. We develop and compare two UDE architectures in a progressive workflow: (A) a dual-network hybrid approach that augments mechanistic Gompertz growth with neural networks to capture unmodeled immune dynamics ($R^2 = 0.897$), and (B) a structured-parameter UDE that uses neural networks to learn context-dependent parameters within classical Gompertz and Michaelis--Menten frameworks ($R^2 = 0.991$). Both variants enforce biological constraints through physics-informed loss functions. Cross-validation with validation-based early stopping yields mean $R^2 = 0.945$ on held-out groups. Our implementation leverages DifferentialEquations.jl, Flux.jl, and SciMLSensitivity.jl. Code and data are publicly available.
 \end{abstract}
 
-\noindent\textbf{Keywords:} Universal Differential Equations; Scientific Machine Learning; Tumor--Immune Dynamics; Neural ODEs; Physics-Informed Neural Networks; Julia; SciML; Gompertz Growth; Michaelis--Menten Kinetics; Biological Constraints; Clinical Oncology; Cross-Validation; Forecasting.
+\noindent\textbf{Keywords:} Universal Differential Equations; Scientific Machine Learning; Tumor--Immune Dynamics; Julia; SciML; Gompertz Growth; Michaelis--Menten Kinetics; Cross-Validation.
 
 % ===========================
 \section{Introduction}
 
[…]
