### 2026-04-07 #204 @editorialbot
[REVIEW]: Biologically Consistent Universal Differential Equations for Tumor–Immune Dynamics: Towards a more interpretable approach
**Submitting author:** @ans036 (<a href="http://orcid.org/0009-0009-7965-7788">Anish  Sarkar</a>)
**Repository:** https://github.com/ans036/tumor-immune-ude
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
