+++ paper/paper.yml
@@ -19,7 +19,7 @@ keywords:
   - Biological Constraints
   - Clinical Oncology
 bibliography: ref.bib
-# Optional metadata JuliaCon supports (safe to keep):
+
 github: "https://github.com/ans036/tumor-ude"
 tags:
   - julia
