### 2026-04-08 e14d0d2
Update paper.tex
