+++ paper/paper.tex
@@ -14,8 +14,8 @@
 
 \graphicspath{{figures/}}
 
-\title{Universal Differential Equations for Tumor--Immune Dynamics:\\
-A Progressive Modeling Workflow in Julia's SciML Ecosystem}
+\title{Biologically Consistent Universal Differential Equations for Tumor--Immune Dynamics:\\
+Towards a more interpretable approach}
 
 \author[1]{Anish Sarkar\thanks{ORCID: \href{https://orcid.org/0009-0009-7965-7788}{0009-0009-7965-7788}}}
 \affil[1]{Netaji Subhash Engineering College (MAKAUT)}
@@ -25,7 +25,7 @@ A Progressive Modeling Workflow in Julia's SciML Ecosystem}
 \maketitle
 
 \begin{abstract}
-We present a study of Universal Differential Equations (UDEs) for modeling tumor--immune dynamics using Julia's SciML ecosystem. We develop and compare two UDE architectures in a progressive workflow: (A) a dual-network hybrid approach that augments mechanistic Gompertz growth with neural networks to capture unmodeled immune dynamics ($R^2 = 0.897$), and (B) a structured-parameter UDE that uses neural networks to learn context-dependent parameters within classical Gompertz and Michaelis--Menten frameworks ($R^2 = 0.991$). Both variants enforce biological constraints through physics-informed loss functions. Cross-validation with validation-based early stopping yields mean $R^2 = 0.945$ on held-out groups. Our implementation leverages DifferentialEquations.jl, Flux.jl, and SciMLSensitivity.jl. Code and data are publicly available.
+We present a study of Universal Differential Equations (UDEs) for modeling tumor--immune dynamics---a progressive modeling workflow in Julia's SciML ecosystem. We develop and compare two UDE architectures: (A) a dual-network hybrid approach that augments mechanistic Gompertz growth with neural networks to capture unmodeled immune dynamics ($R^2 = 0.897$), and (B) a structured-parameter UDE that uses neural networks to learn context-dependent parameters within classical Gompertz and Michaelis--Menten frameworks ($R^2 = 0.991$). Both variants enforce biological constraints through physics-informed loss functions. Cross-validation with validation-based early stopping yields mean $R^2 = 0.945$ on held-out groups. Our implementation leverages DifferentialEquations.jl, Flux.jl, and SciMLSensitivity.jl. Code and data are publicly available.
 \end{abstract}
 
 \noindent\textbf{Keywords:} Universal Differential Equations; Scientific Machine Learning; Tumor--Immune Dynamics; Julia; SciML; Gompertz Growth; Michaelis--Menten Kinetics; Cross-Validation.
@@ -37,15 +37,17 @@ Modeling tumor--immune interactions is a long-standing challenge in computationa
 
 Julia's SciML ecosystem provides strong infrastructure for UDE development through DifferentialEquations.jl \cite{rackauckas2017de}, Flux.jl \cite{innes2018flux}, and SciMLSensitivity.jl for adjoint-based gradient computation \cite{bezanson2017julia,innes2019differentiable,rackauckas2019diffeqflux}. While UDEs have been applied in hydrology and epidemiology \cite{shen2023differentiable,aboelyazeed2023differentiable,dandekar2020machine}, their application to tumor--immune modeling with rigorous cross-validation remains underexplored.
 
-This paper makes the following contributions: (i) a dual-network hybrid UDE that augments Gompertz growth with learned immune dynamics, (ii) a structured-parameter UDE where neural networks learn context-dependent parameters within classical frameworks, (iii) cross-validation with validation-based early stopping for robust evaluation, and (iv) a fully reproducible Julia implementation.
[…]
