### 2026-04-08 #204 @ans036
Hi, I've updated `paper.tex` with the following changes:

1. **Title updated** to "Biologically Consistent Universal Differential Equations for Tumor–Immune Dynamics: Towards a more interpretable approach".

2. **Section heading casing normalized** — e.g. `Software Design` → `Software design`.

However, the editorialbot is still failing to discover the required sections (`Statement of need`, `State of the field`, `Software design`, `Research impact statement`, `AI usage disclosure`) despite all of them being present in the LaTeX source as `\section{...}` headings. The wordcount check succeeds, so the bot is reading the file correctly — it just isn't matching the section titles.

Could this be a parsing issue with the bot's section detection on `.tex` files? Happy to adjust formatting if there's a specific heading convention the bot expects (e.g. specific casing, label format, or section depth).
