### 2026-04-13 383f211
Address reviewer comments (JuliaCon PR #204): regenerate C4-T5 trajectories, merge forecast/metric figures, clarify extrapolation/stabilization, add equation/hyperparameter definitions
- paper/paper.tex: 22 reviewer comments addressed (affiliation, abstract URL, Eq 3 R_kill rename, Eq 4/5/7/9 notation clarifications, extrapolation note, hyperparameter bounds with citations, L-BFGS/patience specifics, seed list, cross-validation identifiers, Phase 1/2 sequential framing, classical-vs-UDE R^2 comparison, Fig 12 -> C4-T5, Fig 9/10 and 13/14 merged, Table 2 variance explanation)
- paper/paper.yml:
[…]
