+++ paper/paper.tex
@@ -4,28 +4,28 @@
 \usepackage{graphicx}
 \usepackage{booktabs}
 \usepackage{amsmath,amssymb}
-\usepackage{siunitx}
+% \usepackage{siunitx} (removed - not used)
 \usepackage{hyperref}
 \usepackage{enumitem}
 \usepackage{float}
 \usepackage{microtype}
-\usepackage{algorithm}
-\usepackage{algorithmic}
+% \usepackage{algorithm}
+% \usepackage{algorithmic}
 
 \graphicspath{{figures/}}
 
 \title{Biologically Consistent Universal Differential Equations for Tumor--Immune Dynamics:\\
 Towards a more interpretable approach}
 
-\author[1]{Anish Sarkar\thanks{ORCID: \href{https://orcid.org/0009-0009-7965-7788}{0009-0009-7965-7788}}}
-\affil[1]{Netaji Subhash Engineering College (MAKAUT)}
+\author[1]{Anish Sarkar}
+\affil[1]{Netaji Subhash Engineering College (MAKAUT), ORCID: \href{https://orcid.org/0009-0009-7965-7788}{0009-0009-7965-7788}}
 \date{}
 
 \begin{document}
 \maketitle
 
 \begin{abstract}
-We present a study of Universal Differential Equations (UDEs) for modeling tumor--immune dynamics---a progressive modeling workflow in Julia's SciML ecosystem. We develop and compare two UDE architectures: (A) a dual-network hybrid approach that augments mechanistic Gompertz growth with neural networks to capture unmodeled immune dynamics ($R^2 = 0.897$), and (B) a structured-parameter UDE that uses neural networks to learn context-dependent parameters within classical Gompertz and Michaelis--Menten frameworks ($R^2 = 0.991$). Both variants enforce biological constraints through physics-informed loss functions. Cross-validation with validation-based early stopping yields mean $R^2 = 0.945$ on held-out groups. Our implementation leverages DifferentialEquations.jl, Flux.jl, and SciMLSensitivity.jl. Code and data are publicly available.
+We present a study of Universal Differential Equations (UDEs) for modeling tumor--immune dynamics---a progressive modeling workflow in Julia's SciML ecosystem. We develop and compare two UDE architectures: (A) a dual-network hybrid approach that augments mechanistic Gompertz growth with neural networks to capture unmodeled immune dynamics ($R^2 = 0.897$), and (B) a structured-parameter UDE that uses neural networks to learn context-dependent parameters within classical Gompertz and Michaelis--Menten frameworks ($R^2 = 0.991$). Both variants enforce biological constraints through physics-informed loss functions. Cross-validation with validation-based early stopping yields mean $R^2 = 0.945$ on held-out groups. Our implementation leverages DifferentialEquations.jl, Flux.jl, and SciMLSensitivity.jl. Code and data are publicly available at \url{https://github.com/ans036/tumor-immune-ude}.
 \end{abstract}
 
 \noindent\textbf{Keywords:} Universal Differential Equations; Scientific Machine Learning; Tumor--Immune Dynamics; Julia; SciML; Gompertz Growth; Michaelis--Menten Kinetics; Cross-Validation.
@@ -33,7 +33,7 @@ We present a study of Universal Differential Equations (UDEs) for modeling tumor
 % ===========================
 \section{Introduction}
 
[…]
