### 2026-04-13 e5c11f4
Add missing DOIs: bezanson2012julia, raymundo2024comparative (arXiv), ramadhan2024data (MIT handle)
