+++ paper/ref.bib
@@ -119,7 +119,8 @@
   year          = {2012},
   eprint        = {1209.5145},
   archivePrefix = {arXiv},
-  primaryClass  = {cs.PL}
+  primaryClass  = {cs.PL},
+  doi           = {10.48550/arXiv.1209.5145}
 }
 
 @article{lu2021deeponet,
@@ -228,8 +229,10 @@
   title         = {A Comparative Study of Neural{ODE} and Universal Differential Equation Approaches to Solving Chandrasekhar's White Dwarf Equation},
   author        = {Vazquez Martinez, Raymundo and Dandekar, Raj Abhijit and Dandekar, Rajat and Panat, Sreedath},
   year          = {2024},
-  note          = {Preprint},
-  howpublished  = {\url{https://drive.google.com/file/d/1mzlaqPmxwhRNG3s7yiHTCL-hVZXsE6YO/view}}
+  eprint        = {2410.14998},
+  archivePrefix = {arXiv},
+  primaryClass  = {astro-ph.SR},
+  doi           = {10.48550/arXiv.2410.14998}
 }
 
 @article{shen2023differentiable,
@@ -286,7 +289,8 @@
   title  = {Data-driven ocean modeling using neural differential equations},
   author = {Ramadhan, Ali},
   school = {Massachusetts Institute of Technology},
-  year   = {2024}
+  year   = {2024},
+  url    = {https://hdl.handle.net/1721.1/153671}
 }
 
 @article{aboelyazeed2023differentiable,
