### 2026-04-13 4a4893d
Replace Ramadhan thesis (no DOI) with published arXiv paper (10.48550/arXiv.2010.12559)
