+++ paper/ref.bib
@@ -285,12 +285,14 @@
   doi           = {10.48550/arXiv.2310.20079}
 }
 
-@phdthesis{ramadhan2024data,
-  title  = {Data-driven ocean modeling using neural differential equations},
-  author = {Ramadhan, Ali},
-  school = {Massachusetts Institute of Technology},
-  year   = {2024},
-  url    = {https://hdl.handle.net/1721.1/153671}
+@article{ramadhan2024data,
+  title         = {Capturing missing physics in climate model parameterizations using neural differential equations},
+  author        = {Ramadhan, Ali and Marshall, John and Souza, Andre and Lee, Xin Kai and Piterbarg, Ulyana and Hillier, Adeline and Wagner, Gregory LeClaire and Rackauckas, Christopher and Hill, Chris and Campin, Jean-Michel and Ferrari, Raffaele},
+  year          = {2020},
+  eprint        = {2010.12559},
+  archivePrefix = {arXiv},
+  primaryClass  = {physics.ao-ph},
+  doi           = {10.48550/arXiv.2010.12559}
 }
 
 @article{aboelyazeed2023differentiable,
