### 2026-04-13 f18d39f
Fix hybrid parameter claims to match actual learned values from results
