+++ paper/paper.tex
@@ -190,7 +190,7 @@ de Pillis--Radunskaya & Immune + therapy terms & None & 0.845 \\
 
 \subsection{Phase 1: Hybrid UDE}
 
-The hybrid UDE converges stably through both AdamW and L-BFGS phases, achieving $R^2 = 0.897$ (Figure~\ref{fig:hybrid_training_loss}). The Immune Response Network learns biologically meaningful patterns: volume-dependent activation consistent with antigen presentation, saturation at high immune fractions, and threshold effects at low immune levels (Figure~\ref{fig:irn_surface}). The Time-Aware Correction Network captures temporal effects not present in the mechanistic backbone (Figure~\ref{fig:tacn_surface}). While the correction magnitude does increase over time---reflecting growing discrepancy between the Gompertz backbone and observed dynamics---the \emph{rate of increase} in correction diminishes at late times, indicating that the learned correction approaches an asymptotic regime rather than diverging. We use ``stabilization'' to refer to this plateauing behavior of the correction's derivative, not a decrease in absolute magnitude. Learned parameters exhibit group-specific variation that correlates with biological characteristics (Figure~\ref{fig:hybrid_params}): growth rates span $r \in [0.02, 0.45]$~day$^{-1}$, with aggressive tumor scenarios (e.g., T4, T5) learning higher $r$ values than slow-growing ones (T1, T2), and immune strength parameters $\alpha$ are larger for groups with higher measured immune cell counts.
+The hybrid UDE converges stably through both AdamW and L-BFGS phases, achieving $R^2 = 0.897$ (Figure~\ref{fig:hybrid_training_loss}). The Immune Response Network learns biologically meaningful patterns: volume-dependent activation consistent with antigen presentation, saturation at high immune fractions, and threshold effects at low immune levels (Figure~\ref{fig:irn_surface}). The Time-Aware Correction Network captures temporal effects not present in the mechanistic backbone (Figure~\ref{fig:tacn_surface}). While the correction magnitude does increase over time---reflecting growing discrepancy between the Gompertz backbone and observed dynamics---the \emph{rate of increase} in correction diminishes at late times, indicating that the learned correction approaches an asymptotic regime rather than diverging. We use ``stabilization'' to refer to this plateauing behavior of the correction's derivative, not a decrease in absolute magnitude. Learned parameters exhibit group-specific variation (Figure~\ref{fig:hybrid_params}): growth rates span $r \in [0.10, 0.34]$~day$^{-1}$ across groups, with the C4$|$T5 group learning the highest rate ($r = 0.34$~day$^{-1}$) consistent with its rapid observed growth. Immune strength parameters $\alpha$ range from 0.11 to 0.38, showing group-level variation that reflects the interplay between immune cell counts and tumor-specific sensitivity rather than a simple monotonic relationship.
 
 \begin{figure}[H]\centering
 \includegraphics[width=\linewidth]{hybrid_training_loss.png}
