### 2026-04-13 19721f6
Add installation guide (README section + standalone INSTALLATION.md)
