### 2026-04-13 afa6b92
Add CONTRIBUTING.md and docstrings to core functions in both pipelines
