### 2026-04-13 9225621
Fix truncated file paths in script guards (unterminated string literal)
