### 2026-04-22 #204 @agchesebro
@ans036 I have a few comments/suggestions for revisions I can add over in the repo. Do you prefer I open a single issue per comment or one issue with a checklist? Happy to do whatever is better for your workflow :)
