### 2026-04-27 27f8a12
Address reviewer comment #3: data origin/timing, expanded loss section, Fig 12 caption, AdamW budget note, README community guidelines
