+++ paper/paper.tex
@@ -118,16 +118,30 @@ A physics prior aligns learned immune parameters with observed killing rates via
 \mathcal{L}_{\text{physics}} = \lambda_{\text{phys}}\left(1 - \rho_{\text{rank}}(I_{\text{static}}, \hat{c}_{\text{kill}})\right)^2
 \label{eq:physics_prior}
 \end{equation}
-The total loss combines volume fitting, physics priors, negativity penalties, consistency, and $L_2$ regularization:
+The Phase~1 (hybrid UDE) total loss combines volume fitting, derivative matching, a smoothness penalty, physics priors, negativity penalties, and $L_2$ regularization:
 \begin{align}
-\mathcal{L} &= \lambda_{\text{vol}} \mathcal{L}_{\text{volume}} + \lambda_{\text{phys}} \mathcal{L}_{\text{physics}} + \lambda_{\text{neg}} \mathcal{L}_{\text{negativity}} + \lambda_{\text{cons}} \mathcal{L}_{\text{consistency}} + \lambda_{\text{reg}} \mathcal{L}_{\text{reg}}
+\mathcal{L}^{(1)} &= w_{\text{vol}}\,\lambda_{\text{vol}} \mathcal{L}_{\text{volume}}
++ w_{\text{deriv}}\,\lambda_{\text{deriv}} \mathcal{L}_{\text{deriv}} \nonumber\\
+&\quad + \lambda_{\text{smooth}} \mathcal{L}_{\text{smooth}}
++ \lambda_{\text{phys}} \mathcal{L}_{\text{physics}} \nonumber\\
+&\quad + \lambda_{\text{neg}} \mathcal{L}_{\text{negativity}} + \lambda_{\text{reg}} \mathcal{L}_{\text{reg}}
 \label{eq:total_loss}
 \end{align}
-where $\mathcal{L}_{\text{volume}} = \frac{1}{N}\sum_{g,j}(\hat{V}_g(t_{g,j}) - V^{\text{obs}}_g(t_{g,j}))^2$ is the mean squared error between predicted and observed tumor volumes across all $N$ data points, $\mathcal{L}_{\text{negativity}} = \sum_{g,j}\max(0, -\hat{V}_g(t_{g,j}))^2$ penalizes non-physical negative volume predictions, and $\mathcal{L}_{\text{reg}} = \|\boldsymbol{\theta}_{\text{nn}}\|_2^2$ is $L_2$ weight decay on neural network parameters. The weights $\lambda_{\text{vol}}=1.0$, $\lambda_{\text{phys}}=0.1$, $\lambda_{\text{neg}}=10^{-3}$, $\lambda_{\text{cons}}=0.5$, and $\lambda_{\text{reg}}=10^{-5}$ were selected by grid search on training loss convergence and biological consistency.
+where the individual terms are:
+\begin{itemize}[leftmargin=*]
+  \item $\mathcal{L}_{\text{volume}} = \frac{1}{N}\sum_{g,j}(\hat{V}_g(t_{g,j}) - V^{\text{obs}}_g(t_{g,j}))^2$: mean squared error between predicted and observed tumor volumes ($N$ = total data points across groups);
+  \item $\mathcal{L}_{\text{deriv}} = \frac{1}{M}\sum_{g,j}\bigl(\dot{\hat{V}}_g(t_{g,j}) - \dot{V}^{\text{obs}}_g(t_{g,j})\bigr)^2$: derivative-matching loss, where $\dot{V}^{\text{obs}}$ is estimated from observations by central finite differences and $\dot{\hat{V}}$ is the model RHS evaluated at the predicted state. This term encourages the learned dynamics, not just the integrated trajectory, to match the data;
[…]
