### 2026-04-27 d886755
Attribute dataset to Atsou et al. (kevin-atsou/tumorgrowth, CC0); add citations to ref.bib, paper, README, and Data/README
