+++ paper/paper.tex
@@ -159,10 +159,13 @@ We use Tsit5() with InterpolatingAdjoint() for gradient computation, with tolera
 \section{Experimental Setup}
 
 \subsection{Data}
+\label{sec:data}
 
-The temporal dataset (\texttt{tumor\_time\_to\_event\_data.csv}) contains longitudinal tumor-volume measurements from a controlled in~vitro tumor--immune co-culture experiment, in which tumor cells were grown under three distinct levels of administered immune (effector) cell concentration (C2, C3, C4) and across five tumor replicates/scenarios (T1--T5). For each (KineticID, TumorID) pair the dataset records: discrete observation times (in days), the measured tumor volume (mm\textsuperscript{3}), and the concurrently measured immune cell count. Measurements were taken at days $t \in \{14, 15, 21, 22, 28, 29, 36\}$, i.e.\ approximately weekly with a paired observation each week, giving up to 7 time points per group (some groups have fewer due to missing or excluded measurements). A static dataset (\texttt{tumor\_volume\_vs\_Im\_cells\_rate.csv}) provides 39 paired measurements of tumor volume and immune killing rate (rank correlation $r = -0.82$), used for physics-informed constraints. Both CSVs are included in the repository under \texttt{Data/}; they are derived from the experimental records of the present study and are not drawn from a previously published dataset.
+We use the publicly available murine tumor--immune dataset released by Atsou et~al.\ \cite{atsou2021size,atsou2022equilibrium} alongside their size- and space-structured tumor--immune interaction model. The data are distributed under a Creative Commons Zero (CC0) license at \url{https://github.com/kevin-atsou/tumorgrowth/tree/main/data} and originate from in~vivo experiments using the mSCC38 squamous cell carcinoma cell line in FVB/N wild-type mice, described in detail in \cite{atsou2022equilibrium} (Materials, Section~2.3). We obtained both CSVs unmodified from that repository and re-host them at \texttt{Data/} in our own repository for convenience.
 
-The reported forecast windows in Section~\ref{sec:results} (e.g.\ days 15--36 in Figures~\ref{fig:c4t5_forecast}--\ref{fig:c4t5_forecasts_combined}) correspond to the empirical observation interval of this dataset; the lower bound is the first observation time and the upper bound is the last. Tumor volumes are normalized by 1000~mm\textsuperscript{3} for numerical stability. Sparse immune measurements are interpolated using cubic splines. Data is grouped by \{KineticID, TumorID\}, excluding groups with fewer than 2 time points.
+The temporal dataset (\texttt{tumor\_time\_to\_event\_data.csv}) contains longitudinal tumor-volume measurements grouped by (KineticID, TumorID) pairs, where the KineticID values \{C2, C3, C4\} index three experimental immune-cell conditions and the TumorID values \{T1, \dots, T5\} index five tumor replicates / scenarios. For each pair the file records: observation time (days), tumor volume (mm\textsuperscript{3}), and the concurrently measured immune cell count. Measurements were taken at days $t \in \{14, 15, 21, 22, 28, 29, 36\}$, giving up to 7 time points per group (some groups have fewer due to missing or excluded measurements). A static dataset (\texttt{tumor\_volume\_vs\_Im\_cells\_rate.csv}) provides 39 paired measurements of tumor volume and immune killing rate (rank correlation $r = -0.82$), used for physics-informed constraints. The reported forecast windows in Section~\ref{sec:results} (e.g.\ days 15--36 in Figures~\ref{fig:c4t5_forecast}--\ref{fig:c4t5_forecasts_combined}) correspond to the empirical observation interval of this dataset; the lower bound is the first observation time and the upper bound is the last.
+
[…]
