+++ paper/ref.bib
@@ -295,6 +295,27 @@
   doi           = {10.48550/arXiv.2010.12559}
 }
 
+@article{atsou2021size,
+  title   = {A size and space structured model of tumor growth describes a key role for protumor immune cells in breaking equilibrium states in tumorigenesis},
+  author  = {Atsou, Kevin and Anjuere, Fabienne and Braud, Veronique M. and Goudon, Thierry},
+  journal = {PLOS ONE},
+  volume  = {16},
+  number  = {11},
+  pages   = {e0259291},
+  year    = {2021},
+  doi     = {10.1371/journal.pone.0259291}
+}
+
+@article{atsou2022equilibrium,
+  title   = {Analysis of the equilibrium phase in immune-controlled tumors provides hints for designing better strategies for cancer treatment},
+  author  = {Atsou, Kevin and Khou, Sokchea and Anjuere, Fabienne and Braud, Veronique M. and Goudon, Thierry},
+  journal = {Frontiers in Oncology},
+  volume  = {12},
+  pages   = {878827},
+  year    = {2022},
+  doi     = {10.3389/fonc.2022.878827}
+}
+
 @article{aboelyazeed2023differentiable,
   title   = {A differentiable, physics-informed ecosystem modeling and learning framework for large-scale inverse problems: Demonstration with photosynthesis simulations},
   author  = {Aboelyazeed, Doaa and Xu, Chonggang and Hoffman, Forrest M. and Liu, Jiangtao and Jones, Alex W. and Rackauckas, Chris and Lawson, Kathryn and Shen, Chaopeng},
