### 2026-04-27 b8b9fc0
Add CITATION.cff and prominent Citation/badges/paper-link section to README; fix repo layout listing
