+++ README.md
@@ -1,5 +1,15 @@
 # Tumor-UDE: Physics-Informed & Pure-Mechanistic UDEs for Tumor–Immune Dynamics
 
+[![CI](https://github.com/ans036/tumor-immune-ude/actions/workflows/ci.yml/badge.svg)](https://github.com/ans036/tumor-immune-ude/actions/workflows/ci.yml)
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+[![Julia ≥ 1.9](https://img.shields.io/badge/Julia-%E2%89%A5%201.9-purple.svg)](https://julialang.org/)
+[![JuliaCon Proceedings](https://img.shields.io/badge/JuliaCon-Proceedings%20Review%20%23204-blue)](https://github.com/JuliaCon/proceedings-review/issues/204)
+[![DOI Pending](https://img.shields.io/badge/DOI-pending-lightgrey)](https://github.com/JuliaCon/proceedings-review/issues/204)
+
+> **Companion code for the JuliaCon Proceedings submission**
+> *Biologically Consistent Universal Differential Equations for Tumor–Immune Dynamics: Towards a More Interpretable Approach* — Anish Sarkar (2026).
+> Manuscript and figures: [`paper/`](paper/) · Reviewer proceedings issue: [JuliaCon/proceedings-review#204](https://github.com/JuliaCon/proceedings-review/issues/204) · Citation file: [`CITATION.cff`](CITATION.cff)
+
 This repository contains two Julia pipelines for modeling tumor–immune dynamics from time-series tumor volume and immune measurements:
 
 - **Classic (Enhanced PI-NODE)** — Physics-informed neural ODE with residual/correction subnetworks and multi-restart training (`src/JuliaconSubmission.jl`).
@@ -14,13 +24,28 @@ Both pipelines output training curves, observed vs predicted trajectories, count
 ```
 .
 ├── src/
-│   ├── JuliaconSubmission.jl      # Classic: enhanced PI-NODE pipeline
-│   └── FutureWorks.jl             # Future works: pure-mechanistic UDE
+│   ├── JuliaconSubmission.jl      # Phase 1: enhanced PI-NODE / hybrid UDE pipeline
+│   ├── FutureWorks.jl             # Phase 2: pure-mechanistic / structured-parameter UDE
+│   └── Project.toml               # Julia environment (deps for both pipelines)
 ├── test/
-│   ├── runtests.jl                # Classic smoke/basics
-│   └── pure_mechanistic_tests.jl  # Future-works smoke/basics
+│   ├── runtests.jl                # Phase 1 smoke/basics
+│   └── pure_mechanistic_tests.jl  # Phase 2 smoke/basics
+├── Data/
+│   ├── tumor_time_to_event_data.csv         # Longitudinal tumor + immune measurements
+│   ├── tumor_volume_vs_Im_cells_rate.csv    # Static volume vs killing-rate pairs
+│   └── README.md                            # Data attribution, citations, column key
+├── paper/
+│   ├── paper.tex                  # JuliaCon proceedings manuscript
+│   ├── paper.yml                  # JuliaCon metadata
+│   ├── ref.bib                    # Bibliography
+│   └── figures/                   # Publication figures
+├── enhanced_results_2025-07-20_161123/      # Saved Phase 1 example outputs
+├── results_PureMechanisticUDE_2025-07-19_112322/  # Saved Phase 2 example outputs
 ├── .github/workflows/ci.yml       # Matrix CI (classic & pure) + artifact upload
-└── results/                       # Created at runtime (plots & analysis)
+├── CITATION.cff                   # Machine-readable citation metadata
+├── CONTRIBUTING.md                # Contribution workflow
[…]
