### 2026-08-04 aeb2be2
Address reviewer follow-up (issue #2): parameter-count comparison + learned immune-term interpretation
Adds the two analyses the reviewer requested to substantiate how UDEs address
the stated limitations of classical models:

1. Parameter Count subsection (Sec. Results): exact counts extracted from the
   real network objects -- Kuznetsov-Taylor (8), de Pillis-Radunskaya (~20),
   Hybrid UDE Phase 1 (10,275), Structured-Param UDE Phase 2 (5,476) -- and an
   honest explanation that the UDE's advantage is eliminating the prescribed
   interaction form and manual calibration, not raw parameter p
[…]
