+++ paper/paper.tex
@@ -79,7 +79,7 @@ where $R_{\text{kill}}$ denotes the immune killing rate, $I(t)$ is immune cell c
 
 \subsection{Phase 1: Dual-Network Hybrid UDE}
 
-The Gompertz model (Eq.~\ref{eq:gompertz}) describes tumor growth in isolation but does not account for immune cell interactions. Classical models like Kuznetsov--Taylor \cite{kuznetsov1994} address this through explicit effector-cell compartments, but require many parameters and strong mechanistic assumptions. Our hybrid approach instead uses neural networks to \emph{learn} the immune interaction and residual temporal dynamics directly from data, while retaining the well-established Gompertz backbone. This avoids prescribing the functional form of immune suppression a priori, allowing the data to reveal which interactions matter. The hybrid UDE augments Gompertz growth with two neural networks:
+The Gompertz model (Eq.~\ref{eq:gompertz}) describes tumor growth in isolation but does not account for immune cell interactions. Classical models like Kuznetsov--Taylor \cite{kuznetsov1994} address this through explicit effector-cell compartments, but require the modeler to prescribe the functional form of the immune interaction and to calibrate its biologically named parameters---several of which are difficult to identify from sparse data (we quantify this trade-off in Section~\ref{sec:param-count}). Our hybrid approach instead uses neural networks to \emph{learn} the immune interaction and residual temporal dynamics directly from data, while retaining the well-established Gompertz backbone. This avoids prescribing the functional form of immune suppression a priori, allowing the data to reveal which interactions matter. The hybrid UDE augments Gompertz growth with two neural networks:
 \begin{align}
 \frac{dV}{dt} &= \underbrace{r_g V \ln\left(\frac{K_g}{V}\right)}_{\text{Gompertz}} - \underbrace{\alpha_g \phi_{\text{IRN}}(V_n, I_n)}_{\text{Learned Immune}} + \underbrace{\beta w(V_n) \phi_{\text{TACN}}(V_n, t_n)}_{\text{Temporal Correction}}
 \label{eq:dual_network_ude}
@@ -206,6 +206,36 @@ de Pillis--Radunskaya & Immune + therapy terms & None & 0.845 \\
 \label{tab:model_comparison}
 \end{table*}
 
+\subsection{Parameter Count and the Nature of Model Complexity}
+\label{sec:param-count}
+
+A recurring motivation for hybrid modeling is that classical mechanistic models are ``parameter-heavy.'' We make this comparison explicit and, in doing so, clarify what it does and does not mean here. Table~\ref{tab:param_counts} reports exact counts; the UDE figures were extracted directly from the model objects (\texttt{length(Flux.destructure(nn)[1])}), and the classical figures from the reference model definitions.
+
+The Kuznetsov--Taylor model is a two-compartment ODE system (effector cells $E$, tumor cells $T$) with eight kinetic parameters $\{s,p,g,m,d,a,b,n\}$ (BioModels \texttt{BIOMD0000000762}) \cite{kuznetsov1994}; the de Pillis--Radunskaya system adds NK-cell, CD8\textsuperscript{+}, and circulating-lymphocyte compartments together with therapy terms, on the order of twenty parameters \cite{depillis2005validated}. By contrast, our structured-parameter UDE (Phase~2) contains a single shared network of $5{,}476$ weights, and the dual-network hybrid (Phase~1) contains $10{,}242$ network weights plus three scalar parameters for each of the eleven groups it retains.
+
+\begin{table}[H]\centering
+\small
+\begin{tabular}{@{}lcccc@{}}
+\toprule
+\textbf{Model} & \textbf{States} & \textbf{Mech.} & \textbf{NN} & \textbf{Total} \\
+ & & \textbf{params} & \textbf{weights} & \\
+\midrule
[…]
