+++ results/immune_interpretation/final/immune_summary.txt
@@ -1,0 +1,20 @@
+checkpoint=D:/tumor-immune-ude-push/results/immune_interpretation/phase2_lbfgs.jld2
+global_R2=0.928269
+monotonic_in_I=true
+R_at_Imin=0.003718
+R_at_Imax=0.283701
+R_at_Vmin=0.004495
+R_at_Vmax=0.174098
+Vhalf_norm=0.256248
+Vhalf_mm3=256.248
+c_kill_min=6.1448
+c_kill_med=8.2185
+c_kill_max=9.3770
+h_sat_min=0.8989
+h_sat_med=1.3667
+h_sat_max=2.5088
+h_sat_mm3_min=898.9
+h_sat_mm3_max=2508.8
+spearman_ckill_vs_static=0.8008
+median_suppression_ratio=0.8717
+max_suppression_ratio=3.0503

[…]
