### 2026-08-04 d271a18
Re-render all figures at large fonts for print readability (reviewer figure item)
Re-rendered every paper figure with enlarged fonts so they are legible at
publication size:
- Phase 2 (group dynamics, training loss, parameter trajectories) via the
  original create_all_plots with a large-font theme, preserving all detail
  (High Immune Activity markers, etc.); component-dynamics legend moved outside
  the axes to avoid overlapping the curves.
- Phase 1 (training loss, learned parameter distributions, and the Immune
  Response and Time-Aware Correction network surfaces) from a fresh bounded
  P
[…]
