+++ paper/paper.tex
@@ -238,7 +238,7 @@ The UDE thus does not reduce the raw parameter count; it reduces the \emph{model
 
 \subsection{Phase 1: Hybrid UDE}
 
-The hybrid UDE converges stably through both AdamW and L-BFGS phases, achieving $R^2 = 0.897$ (Figure~\ref{fig:hybrid_training_loss}). The Immune Response Network learns biologically meaningful patterns: volume-dependent activation consistent with antigen presentation, saturation at high immune fractions, and threshold effects at low immune levels (Figure~\ref{fig:irn_surface}). The Time-Aware Correction Network captures temporal effects not present in the mechanistic backbone (Figure~\ref{fig:tacn_surface}). While the correction magnitude does increase over time---reflecting growing discrepancy between the Gompertz backbone and observed dynamics---the \emph{rate of increase} in correction diminishes at late times, indicating that the learned correction approaches an asymptotic regime rather than diverging. We use ``stabilization'' to refer to this plateauing behavior of the correction's derivative, not a decrease in absolute magnitude. Learned parameters exhibit group-specific variation (Figure~\ref{fig:hybrid_params}): growth rates span $r \in [0.10, 0.34]$~day$^{-1}$ across groups, with the C4$|$T5 group learning the highest rate ($r = 0.34$~day$^{-1}$) consistent with its rapid observed growth. Immune strength parameters $\alpha$ range from 0.11 to 0.38, showing group-level variation that reflects the interplay between immune cell counts and tumor-specific sensitivity rather than a simple monotonic relationship.
+The hybrid UDE converges stably through both AdamW and L-BFGS phases, achieving $R^2 = 0.897$ (Figure~\ref{fig:hybrid_training_loss}). The Immune Response Network learns biologically meaningful patterns: volume-dependent activation consistent with antigen presentation, saturation at high immune fractions, and threshold effects at low immune levels (Figure~\ref{fig:irn_surface}). The Time-Aware Correction Network captures temporal effects not present in the mechanistic backbone (Figure~\ref{fig:tacn_surface}). While the correction magnitude does increase over time---reflecting growing discrepancy between the Gompertz backbone and observed dynamics---the \emph{rate of increase} in correction diminishes at late times, indicating that the learned correction approaches an asymptotic regime rather than diverging. We use ``stabilization'' to refer to this plateauing behavior of the correction's derivative, not a decrease in absolute magnitude. Learned parameters exhibit group-specific variation (Figure~\ref{fig:hybrid_params}): growth rates span $r \in [0.07, 0.18]$~day$^{-1}$ across groups, with the fastest-growing tumor scenarios learning the highest rates, consistent with their rapid observed growth. Immune strength parameters $\alpha$ range from 0.13 to 0.39, showing group-level variation that reflects the interplay between immune cell counts and tumor-specific sensitivity rather than a simple monotonic relationship.
 
 \begin{figure}[H]\centering
 \includegraphics[width=\linewidth]{hybrid_training_loss.png}
@@ -266,7 +266,7 @@ The hybrid UDE converges stably through both AdamW and L-BFGS phases, achieving 
 
 \subsection{Phase 2: Structured-Parameter UDE}
 
[…]
