### 2026-08-05 127e75d
Fix paper framing per reviewer: remove self-contradictory motivations
Addresses the reviewer's core scientific objection that the paper's stated
motivation contradicted its own content:
- 'State of the field' no longer implies classical models lack immune terms
  (de Pillis-Radunskaya and Kuznetsov-Taylor DO model immunity explicitly);
  reframes their difficulty as prescribing the interaction FORM and estimating
  unidentifiable constants, not parameter count.
- Removes 'parameter-heavy' as a motivation for UDEs and states explicitly that
  the UDE carries FAR MORE parameters than classical m
[…]
