+++ paper/paper.tex
@@ -33,7 +33,7 @@ We present a study of Universal Differential Equations (UDEs) for modeling tumor
 % ===========================
 \section{Introduction}
 
-Modeling tumor--immune interactions is a long-standing challenge in computational biology \cite{kuznetsov1994,depillis2005validated,vaghi2020gompertz}. Traditional mechanistic models offer interpretability but often underfit complex dynamics, while pure machine learning approaches lack biological constraints. Universal Differential Equations (UDEs) combine mechanistic structure with neural network flexibility, embedding trainable components within differential equation frameworks to learn missing dynamics while preserving known physics \cite{rackauckas2020ude,karniadakis2021physics}.
+Modeling tumor--immune interactions is a long-standing challenge in computational biology \cite{kuznetsov1994,depillis2005validated,vaghi2020gompertz}. Traditional mechanistic models offer interpretability but require the modeler to prescribe the functional form of each interaction in advance, while pure machine learning approaches lack biological constraints. Universal Differential Equations (UDEs) combine mechanistic structure with neural network flexibility, embedding trainable components within differential equation frameworks to learn missing dynamics while preserving known physics \cite{rackauckas2020ude,karniadakis2021physics}.
 
 Julia's SciML ecosystem provides strong infrastructure for UDE development through DifferentialEquations.jl \cite{rackauckas2017de}, Flux.jl \cite{innes2018flux}, and SciMLSensitivity.jl for adjoint-based gradient computation \cite{bezanson2017julia,innes2019differentiable,rackauckas2019diffeqflux}. While UDEs have been applied in hydrology and epidemiology \cite{shen2023differentiable,aboelyazeed2023differentiable,dandekar2020machine}, their application to tumor--immune modeling with rigorous cross-validation remains underexplored.
 
@@ -43,13 +43,13 @@ This paper presents a progressive modeling workflow for tumor--immune dynamics i
 \section{Statement of need}
 \label{sec:statement-of-need}
 
-Researchers modeling tumor--immune interactions need tools that balance mechanistic interpretability with predictive accuracy. Existing approaches are either purely mechanistic models that underfit or black-box methods that lack biological grounding. Our Julia package provides a UDE-based framework that integrates domain knowledge through physics-informed constraints, learns interpretable biological parameters, and provides cross-validated forecasting---all within Julia's SciML ecosystem.
+Researchers modeling tumor--immune interactions need tools that balance mechanistic interpretability with predictive accuracy. Existing approaches are either purely mechanistic models whose interaction terms must be specified and calibrated by hand, or black-box methods that lack biological grounding. Our Julia package provides a UDE-based framework that integrates domain knowledge through physics-informed constraints, learns the immune interaction from data as interpretable context-dependent parameters, and provides cross-validated forecasting---all within Julia's SciML ecosystem. We are explicit that the contribution is a reproducible, interpretable modeling \emph{workflow}, not a reduction in parameter count relative to classical models.
 
 % ===========================
 \section{State of the field}
 \label{sec:state-of-the-field}
 
[…]
