### 2026-08-05 e62093e
Fix garbled x-axis labels in hybrid_param_distributions figure
The 3-panel histogram had overlapping/cramped x-tick labels (e.g. r-panel
ticks ran together). Re-plotted from the saved Phase-1 checkpoint with fewer,
well-spaced ticks, larger inter-panel margins, and a taller canvas. Curves
(learned r, K, alpha distributions) unchanged.
