### 2026-08-05 563a043
Fix clipped Fig. 12 caption: float placement + shorter caption
parameter_trajectories_C4T5 was forced with [H] at a page bottom; being a tall
4-panel figure with a long caption, the caption overflowed the page and was
clipped. Changed placement to [tbp] so LaTeX positions the full figure+caption
where it fits, and trimmed the caption to two lines.
